Refactored weather-minor css changes
