+++ index.html
@@ -52,8 +52,8 @@
 
             </div>
         </div>
-        <!---Aside--->
-        <div class="aside">
+         <!---Aside--->
+        <aside class="aside">
             <div class="itineraryAside">
                 <h2 class="itineraryHeader">YOUR ITINERARY</h2>
                 <form class="form">
@@ -65,16 +65,9 @@
                     <h4><u>Your Park:</u></h4>
                 </div>
 
-    <!---Aside--->
-    <aside>
-        <div class="itinerary">
-            <div>
-            <div id="park-itinerary"></div>
-            <div id="weather">
-            <div id="bizarre-itinerary"></div>
-            <div id="eatery-itinerary"></div>
-            <button class="save">Save Itinerary</button>
-            </div>
+                <div id="weather">
+                    <h4><u>Upcoming Weather Report:</u></h4>
+                </div>
 
                 <div id="eateryAside">
                     <h4><u>Your Eatery:</u></h4>
@@ -92,7 +85,7 @@
 
             </div>
 
-        </div>
+        </aside>
     </div>
     <!---Footer--->
     <div class="footer">
